+++ src/constants.ts
@@ -72,6 +72,12 @@ export const relationships = {
     sourceType: entities.ACCOUNT._type,
     targetType: entities.ROLE._type,
   },
+  ACCOUNT_MANAGES_MOBILE_DEVICE: {
+    _type: 'google_account_manages_mobile_device',
+    _class: RelationshipClass.MANAGES,
+    sourceType: entities.ACCOUNT._type,
+    targetType: entities.MOBILE_DEVICE._type,
+  },
   SITE_HOSTS_USER: {
     _type: 'google_site_has_user',
     // TODO: Change to HOSTS